Move cluster enrichment into config
